Use `__int128` only when `__SIZEOF_INT128__ == 16`
"__int128" may only be used when targetting a 64-bit platform.
Previously, this type would be used as the longest integer type even
when that condition is not met.

If the type is still preset for compiler that does not define this
macro, it can be freely set by the user.

This change was tested with i686 GCC and x86 MSVC.